[ZVOOQ-7245] directory renaming
