--- payturetypes/__init__.py
@@ -1,12 +1,0 @@
-from . import constants 
-from . import paytureresponse
-from . import cardinfo
-from . import encodedata
-from . import transaction 
-from . import digitalwallet
-from . import api
-from . import ewallet
-from . import inpay
-from . import merchant
-
-
